requirements.txt
python-telegram-bot==20.6
SQLAlchemy==2.0.23
python-dotenv==1.0.1
mysql-connector-python==8.3.0
requests==2.31.0
